+++ ReleaseNotes.md
@@ -20,6 +20,8 @@ All packages were updated to latest versions of their dependencies.
 ### ```ConnectionManager```
 
 - [New] (Preliminary) Added Async support for query methods.
+
+- [Breaking Change] Removed default constructor.
 - [Breaking Change] Changed return type of ```ExecuteQueryWithResults``` from ```T[]``` to ```IEnumerable<T>```.
 - [Breaking Change] Changed return type of ```QueryParameters``` from ```DataParameter[]``` to ```IEnumerable<DataParameter>```.
 - [Breaking Change] Moved ```WithParameters``` to ```DataCommandExtensions``` as an extension method so it returns the correct type of the command.
@@ -48,6 +50,12 @@ is analyzed to determine if it is a connection string or name. When specifying a
 - [New] Added ```GetDateValueOrDefault``` and ```TryGetDateValue``` methods to work with ```Date``` values.
 - [New] Added ```TryGetDateTimeValue``` to be symmetric with other types.
 
+## P3Net.Kraken.Data.Sql
+
+### ```SqlConnectionManager```
+
+- [Breaking Change] Removed default constructor.
+ 
 ## P3Net.Kraken.Linq [New]
 
 Added new namespace and ```LinqExtensions``` to include some asynchronous LINQ commands. The separate namespace prevents collisions with other implementations (such as Entity Framework).